requirements.dev.txt
docker
pytest-flask
inmanta-dev-dependencies[module]
types-PyYAML
types-requests
